refactor: remove duplicated state and chain classes
