+++ AvaliacaoIII-Base/src/br/ifba/edu/aval/AppAvaliacaoBase.java
@@ -3,7 +3,7 @@ package br.ifba.edu.aval;
 import java.time.Duration;
 
 import br.ifba.edu.aval.exception.AtividadeNaoPermitidaException;
-import br.ifba.edu.aval.model.BoletimProva;
+import br.ifba.edu.aval3.state.BoletimProva;
 import br.ifba.edu.aval.model.Dificuldade;
 import br.ifba.edu.aval.model.Idade;
 import br.ifba.edu.aval.model.Mapa;
