--- AvaliacaoIII-Base/src/br/ifba/edu/aval/model/PosProvaState.java
@@ -1,17 +1,0 @@
-package br.ifba.edu.aval.model;
-
-public class PosProvaState extends FaseProvaAdapter {
-    public PosProvaState(BoletimProva boletim) {
-        super(boletim);
-    }
-
-    @Override
-    public void registrarAtrasoPartida(Long minutoPartidaEfetivo) {
-        boletim.minutoPartidaEfetivo = minutoPartidaEfetivo;
-    }
-
-    @Override
-    public Long getMinutosAtraso() {
-        return boletim.minutoPartidaEfetivo - boletim.minutoPartidaPrevisto;
-    }
-}
