--- AvaliacaoIII-Base/src/br/ifba/edu/aval/model/PreProvaState.java
@@ -1,14 +1,0 @@
-package br.ifba.edu.aval.model;
-
-import br.ifba.edu.aval.exception.AtividadeNaoPermitidaException;
-
-public class PreProvaState extends FaseProvaAdapter {
-    public PreProvaState(BoletimProva boletim) {
-        super(boletim);
-    }
-
-    @Override
-    public void apresentarPraLargada() throws AtividadeNaoPermitidaException {
-        boletim.setFase(new MomentoLargadaState(boletim));
-    }
-}
